Unify common_resources and common modules into models folder
