+++ cli/src/pcluster/models/cluster.py
@@ -46,13 +46,16 @@ from pcluster.models.cluster_resources import (
     FiltersParserError,
     ListClusterLogsFiltersParser,
 )
-from pcluster.models.common import BadRequest, Conflict, LimitExceeded, parse_config
-from pcluster.models.common_resources import (
+from pcluster.models.common import (
+    BadRequest,
     CloudWatchLogsExporter,
+    Conflict,
+    LimitExceeded,
     Logs,
     LogStream,
     create_logs_archive,
     export_stack_events,
+    parse_config,
 )
 from pcluster.models.s3_bucket import S3Bucket, S3BucketFactory, S3FileFormat
 from pcluster.schemas.cluster_schema import ClusterSchema
