+++ cli/src/pcluster/models/imagebuilder.py
@@ -46,14 +46,17 @@ from pcluster.constants import (
     STACK_EVENTS_LOG_STREAM_NAME,
 )
 from pcluster.models.cluster_resources import FiltersParserError
-from pcluster.models.common import BadRequest, Conflict, LimitExceeded, parse_config
-from pcluster.models.common_resources import (
+from pcluster.models.common import (
+    BadRequest,
     CloudWatchLogsExporter,
+    Conflict,
+    LimitExceeded,
     LogGroupTimeFiltersParser,
     Logs,
     LogStream,
     create_logs_archive,
     export_stack_events,
+    parse_config,
 )
 from pcluster.models.imagebuilder_resources import (
     BadRequestStackError,
