+++ cli/tests/pcluster/cli/test_get_cluster_log_events.py
@@ -12,7 +12,7 @@ import time
 import pytest
 from assertpy import assert_that
 
-from pcluster.models.common_resources import LogStream
+from pcluster.models.common import LogStream
 
 BASE_COMMAND = ["pcluster", "get-cluster-log-events"]
 REQUIRED_ARGS = {"cluster_name": "clustername", "log_stream_name": "log-stream-name"}
